refactor: single-spa 빌드 설정 통합
