+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "dev": "vite",
     "build": "vite build",
-    "build:single-spa": "vite build --config vite.single-spa.config.ts",
+    "build:single-spa": "vite build --mode single-spa",
     "test": "echo \"Error: no test specified\"",
     "preview": "vite preview",
     "lint": "npx steiger ./src",
